requirements.txt
tensorflow>=1.*
tensorflow-hub
tqdm>=4.*
keras>=2.2.*
numpy>=1.*
# BERT-related dependencies
bert-tensorflow